% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffRgit.R
\name{diffRgit}
\alias{diffRgit}
\title{Create an HTML git diff widget using the diff2html library}
\usage{
diffRgit(diff, inputFormat = "diff", outputFormat = "side-by-side",
  showFiles = TRUE, matching = "lines", width = NULL, height = NULL,
  elementId = NULL)
}
\arguments{
\item{diff}{the result of a 'git diff'}

\item{inputFormat}{'diff' or 'json'}

\item{outputFormat}{'side-by-side' or 'line-by-line'}

\item{showFiles}{boolean to display or not the header with files changed}

\item{matching}{'lines' or 'words'}

\item{width}{width of the widget}

\item{height}{height of the widget}

\item{elementId}{ID of the widget div}
}
\description{
This function create an HTML widget to diplay a graphic view of a 'git diff' result.
}
\note{
All parameters are directly send to diff2html function to be compatible with futures versions.
}
\examples{
library(diffRgit)

#get git diff with system('git diff file1 file2", intern = TRUE) instead
diff <- "diff --git a/README.md b/README.md\\r\\nindex 132c8a2..46909f2 100644\\r\\n--- a/README.md\\r\\n+++ b/README.md\\r\\n@ -98,6 +98,9 @ The HTML output accepts a Javascript object with configuration. Possible options\\r\\n   - `synchronisedScroll`: scroll both panes in side-by-side mode: `true` or `false`, default is `false`\\r\\n   - `matchWordsThreshold`: similarity threshold for word matching, default is 0.25\\r\\n   - `matchingMaxComparisons`: perform at most this much comparisons for line matching a block of changes, default is `2500`\\r\\n+  - `templates`: object with previously compiled templates to replace parts of the html\\r\\n+  - `rawTemplates`: object with raw not compiled templates to replace parts of the html\\r\\n+  > For more information regarding the possible templates look into [src/templates](https://github.com/rtfpessoa/diff2html/tree/master/src/templates)\\r\\n \\r\\n ## Diff2HtmlUI Helper"
diffRgit(diff)

#see diff line by line
diffRgit(diff, outputFormat = 'line-by-line')

#remove header
diffRgit(diff, showFiles = FALSE)

}
\references{
See \url{https://diff2html.xyz} for more informations.
}
