% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffRgit.R
\name{diffRgit}
\alias{diffRgit}
\title{Create an HTML git diff widget using the diff2html library}
\usage{
diffRgit(diff, inputFormat = "diff", outputFormat = "side-by-side",
  showFiles = TRUE, matching = "lines", width = NULL, height = NULL,
  elementId = NULL)
}
\arguments{
\item{diff}{the result of a 'git diff'}

\item{inputFormat}{'diff' or 'json'}

\item{outputFormat}{'side-by-side' or 'line-by-line'}

\item{showFiles}{boolean to display or not the header with files changed}

\item{matching}{'lines' or 'words'}

\item{width}{width of the widget}

\item{height}{height of the widget}

\item{elementId}{ID of the widget div}
}
\description{
This function create an HTML widget to diplay a graphic view of a 'git diff' result.
}
\note{
All parameters are directly send to diff2html function to be compatible with futures versions.
}
\examples{
library(diffRgit)

#get git diff with system('git diff file1 file2", intern = TRUE) instead
l_diff <- "diff --git a/../diff2widget/file1 b/../diff2widget/file2\\nindex ab63d51..eb01961 100644\\n--- a/../diff2widget/file1\\n+++ b/../diff2widget/file2\\n@ -1,5 +1,4 @\\n-hello!\\n+hello world!\\n Where is Bryan?\\n+Bryan is in the kitchen.\\n bye!\\n-\\n-oups!"
diffRgit(l_diff)

#see diff line by line
diffRgit(l_diff, outputFormat = 'line-by-line')

#remove header
diffRgit(l_diff, showFiles = FALSE)

}
\references{
See \url{https://diff2html.xyz} for more informations.
}
